Implemented mask for OrcidID validation

+++ classes/OrcidId.php
@@ -13,26 +13,31 @@
 	$array = json_decode($data, true);
     curl_close($ch);
 	
-	$firstName = $array['orcid-profile']['orcid-bio']['personal-details']['given-names']['value'];
-	$surname = $array['orcid-profile']['orcid-bio']['personal-details']['family-name']['value'];
-	$name = $firstName." ".$surname;
+	if(array_key_exists('error-desc', $array)){
+		echo json_encode(array("error" => $array['error-desc']['value']));
+	} else{
 	
-	$email = "";
-	
-	if(array_key_exists('orcid-profile', $array)){
-		if(array_key_exists('orcid-bio', $array['orcid-profile'])){
-			if(array_key_exists('contact-details', $array['orcid-profile']['orcid-bio'])){
-				if(array_key_exists('email',  $array['orcid-profile']['orcid-bio']['contact-details'])){
-					foreach($array['orcid-profile']['orcid-bio']['contact-details']['email'] as $x => $arr){
-						if($arr['primary']){
-							$email = $arr['value'];
+		$firstName = $array['orcid-profile']['orcid-bio']['personal-details']['given-names']['value'];
+		$surname = $array['orcid-profile']['orcid-bio']['personal-details']['family-name']['value'];
+		$name = $firstName." ".$surname;
+		
+		$email = "";
+		
+		if(array_key_exists('orcid-profile', $array)){
+			if(array_key_exists('orcid-bio', $array['orcid-profile'])){
+				if(array_key_exists('contact-details', $array['orcid-profile']['orcid-bio'])){
+					if(array_key_exists('email',  $array['orcid-profile']['orcid-bio']['contact-details'])){
+						foreach($array['orcid-profile']['orcid-bio']['contact-details']['email'] as $x => $arr){
+							if($arr['primary']){
+								$email = $arr['value'];
+							}
 						}
+						
 					}
-					
 				}
 			}
-		}
-	}	
-	echo json_encode(array("name" => $name, "email" => $email));
+		}	
+		echo json_encode(array("name" => $name, "email" => $email));
+	}
 	//echo "<pre>".print_r($array)."</pre>";
 ?>
